Separate ultragraph tests into multiple files.
