+++ ultragraph/tests/mod.rs
@@ -1,11 +1,18 @@
 // SPDX-License-Identifier: MIT
 // Copyright (c) "2023" . The DeepCausality Authors. All Rights Reserved.
 
-#[cfg(test)]
-mod ultra_graph_with_matrix_storage_tests;
 
 #[cfg(test)]
 mod constructor_tests;
-
 #[cfg(test)]
 mod error_tests;
+#[cfg(test)]
+mod graph_like_tests;
+#[cfg(test)]
+mod graph_storage_tests;
+#[cfg(test)]
+mod graph_root_tests;
+#[cfg(test)]
+mod graph_shortest_path_tests;
+#[cfg(test)]
+mod graph_outgoing_edges_tests;
